Trying to add correct initialisaiton that works statically for the vm for monitors ... no use of hardcoded asUnlocked constants ... now testing to see if this is ok

+++ com.oracle.max.vm.native/substrate/memory.c
@@ -36,12 +36,12 @@
 Address memory_allocate(Size size)
 {
     static int attempts = 0;
-    printf("Attempt %d Allocating %u\n",attempts++,size);
+    printf("Attempt %d Allocating %u\n",attempts++,(unsigned int)size);
     Address mem = (Address) calloc(1, (size_t) size);
     if (mem % sizeof(void *)) {
         log_println("MEMORY ALLOCATED NOT WORD-ALIGNED (size:%d at address:%x, void* size: %d)", size, mem, sizeof(void *));
     }
-    printf("Address %u\n",mem);
+    printf("Address %u\n",(unsigned int)mem);
     return mem;
 }
